+++ app/controllers/api/v1/plans_controller.rb
@@ -105,53 +105,6 @@ module Api
         scheme = IdentifierScheme.by_name(json[:dmp_id][:type]).first
         Identifier.where(value: json[:dmp_id][:identifier], identifier_scheme: scheme).any?
       end
-
-      # Get the Plan's owner
-      def determine_owner(client:, plan:)
-        contact = plan.contributors.find(&:data_curation?)
-        # Use the contact if it was sent in and has an affiliation defined
-        return contact if contact.present? && contact.org.present?
-
-        # If the contact has no affiliation defined, see if they are already a User
-        user = lookup_user(contributor: contact)
-        return user if user.present?
-
-        # Otherwise just return the client
-        client
-      end
-
-      def lookup_user(contributor:)
-        return nil unless contributor.present?
-
-        identifiers = contributor.identifiers.map do |id|
-          { name: id.identifier_scheme&.name, value: id.value }
-        end
-        user = User.from_identifiers(array: identifiers) if identifiers.any?
-        user = User.find_by(email: contributor.email) unless user.present?
-        user
-      end
-
-      # rubocop:disable Metrics/AbcSize
-      def invite_contributor(contributor:)
-        return nil unless contributor.present?
-
-        # If the user was not found, invite them and attach any know identifiers
-        names = contributor.name&.split || ['']
-        firstname = names.length > 1 ? names.first : nil
-        surname = names.length > 1 ? names.last : names.first
-        user = User.invite!({ email: contributor.email,
-                              firstname: firstname,
-                              surname: surname,
-                              org: contributor.org }, client)
-
-        contributor.identifiers.each do |id|
-          user.identifiers << Identifier.new(
-            identifier_scheme: id.identifier_scheme, value: id.value
-          )
-        end
-        user
-      end
-      # rubocop:enable Metrics/AbcSize
     end
   end
 end
